+++ tools/check.sh
@@ -7,5 +7,6 @@ cd "${ROOT}"
 
 tools/update_java_deps.sh
 bazel run //:buildifier_check
+tools/java_format.sh check
 bazel build //...
 bazel test //...