+++ tools/update_java_deps.sh
@@ -28,4 +28,10 @@ java -jar "${JAR}" \
   --config-file third_party/java/dependencies.yml \
   --cache-directory "${CACHE_DIR}" \
   generate
-bazel run //:buildifier -- MODULE.bazel third_party/java/BUILD.bazel
+java -jar "${JAR}" \
+  --directory "${ROOT}" \
+  --config-file tools/java-format/dependencies.yml \
+  --cache-directory "${CACHE_DIR}" \
+  generate
+bazel run //third_party/java:update_depgen_generated_outputs
+bazel run //:buildifier -- MODULE.bazel third_party/java/BUILD.bazel tools/java-format/BUILD.bazel